RONDB-588: Many fixes for Range Partitioning
Changed name of fragmentNoKeyLen to fragId in SCAN_FRAGREQ
Added DEBUG_FRAGID in Dbdict.cpp
Added DEBUG_RANGE_MAP in Dbdict.cpp
Updated DEBUG_RANGE printouts in Dbdict.cpp
Replaced ZNIL with RNIL for fragId in SUMA,TRIX,DBTUX
Fixed a bug in calculating the new range_map in ADD PARTITION
Moved some debug printouts into DEB_FRAGID
Increased size of indexStatFragId to Uint32 to fit RNIL
Added setStartFidOffset for ERROR_INSERT to set first fragId to 65532
Added a bit of thrjam's for debugging crashes
Ensured that Range partitioned tables avoid single partition pruning
until it is supported by disabling distribution key flag.
Put DICT_TAB_INFO debugging under new debug flag DEBUG_DICT_TABINFO
which is off by default.
Added DEBUG_FRAGID in Trix.cpp
Removed some dead code under trix_index_stat_rep_to_tux_instance
Added a bit of debugging to NdbScanOperation.cpp

+++ storage/ndb/src/kernel/blocks/dbdict/Dbdict.hpp
@@ -482,7 +482,7 @@ class Dbdict : public SimulatedBlock {
      * down, no automatic stats update takes place.  This is not
      * critical and is not worth fixing.
      */
-    Uint16 indexStatFragId;
+    Uint32 indexStatFragId;
     Uint16 indexStatNodes[MAX_REPLICAS];
 
     // pending background request (IndexStatRep::RequestType)
